moved tools_constants.ts to js/constants folder
