+++ static/js/constants/tools_constants.ts
